Rework of structure from talk with Gregorz.
Features --> A feature is a container for a views that make up a feature in the app. It can consist of many views, models and viewmodels, but it should not contain reusable components:
- LocationAndTripSlider
- AddTrip
- AddLocation
- CurrentLocation
- Location
- Trip

Components --> The views that represent components that can be reused in any view. e.g wrappers around existing components like RoundButton, VerticalWeatherView, HorizontalWeatherView etc.